requirements.txt
aiogram>=3.0.0
aiohttp
aiosqlite
matplotlib
numpy
apscheduler
python-dotenv
